alterdec: avoid fs.seek(3) for COptimizedObs
I'm not sure why it's there in the first place but I feel like
explicitly checking for classes for which fs.seek(3) actually does
something useful (or seeking in the class' Read method instead)
would've made more sense. These 3 bytes are essential in case of
COptimizedObs so I'd imagine in many cases the data in these 3
bytes would be absolutely necessary for making sense of the rest
of the object.